refactor: db files structure
