+++ src/back-ws/db/db.ts
@@ -1,33 +1,4 @@
-import { InMemoryRepository } from './memory-repository.js';
+import { userRepository, IUser } from "./users.js";
+import { roomRepository, IRoom } from "./rooms.js";
 
-export interface IUser {
-  id: number;
-  name: string;
-  password: string;
-  wsId: string
-}
-
-export interface IRoom {
-  id: number;
-  roomUsers: {
-    name: string;
-    index: number;
-  }[]
-}
-
-class UserRepository extends InMemoryRepository<IUser, Omit<IUser, 'id'>> {
-  constructor() {
-    super();
-  }
-
-  getByName(name: string) {
-    return this.entities.find(user => user.name === name);
-  }
-
-  getByWsId(wsId: string) {
-    return this.entities.find(user => user.wsId === wsId);
-  }
-}
-
-export const userRepository = new UserRepository();
-export const roomRepository = new InMemoryRepository<IRoom, Omit<IRoom, 'id'>>();
+export { userRepository, IUser, roomRepository, IRoom }
